Completed exercises 4.15 - 4.23 of Blog_lis and Finished Part 4

+++ Part3/Blog_list/controllers/blogs.js
@@ -1,28 +1,64 @@
 const blogsRouter = require('express').Router()
 const { request, response } = require('../app')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
-blogsRouter.get('/', (request, response) => {
-  Blog.find({}).then((blogs) => {
-    response.json(blogs)
-  })
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
+
+  response.json(blogs)
 })
 
-blogsRouter.post('/', (request, response) => {
-  const res = request.body
-  if (res.url == undefined || res.title == undefined){
-    response.status(400).json("url or title property missing!")
-  }
-  else {
-    const blog = new Blog(request.body)
 
-    blog.save().then((result) => {
-      response.status(201).json(result)
-    })
-  }
-})
 
-blogsRouter.delete('/:id' , async (request , response) => {
+blogsRouter.post('/', userExtractor , async (request, response) => {
+  const res = request.body;
+  const user = request.user
+
+
+  if (!user) {
+    return response.status(400).json({ 
+      error: 'No users exist in the database to associate this blog with' 
+    });
+  }
+
+  if (res.url === undefined || res.title === undefined || res.author == undefined) {
+    return response.status(400).json({ error: 'url, title or the author property can not be empty!' });
+  } 
+  
+  const blog = new Blog({
+    title: res.title,
+    author: res.author,
+    url: res.url,
+    likes: res.likes,
+    user: user._id 
+  });
+
+  const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  response.status(201).json(savedBlog);
+});
+
+blogsRouter.delete('/:id' , userExtractor , async (request , response) => {
+
+  const user = request.user
+
+  
+  const blog = await Blog.findById(request.params.id)
+  const userId = user._id
+  const blogPosterId = blog.user.toString()
+
+  if (!(userId == blogPosterId)) {
+    return response.status(400).json({ 
+      error: 'No users exist in the database with this id' 
+    });
+  }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
   })
